backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sinas"
version = "1.0.0"
description = "SINAS - AI Agent Platform with LLM chat, webhooks, MCP tools, and function execution"
authors = [
    {name = "Pulsr One"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "croniter>=1.4.0",
    "asyncpg>=0.29.0",
    "astor>=0.8.1",
    "jsonschema>=4.20.0",
    "python-jose[cryptography]>=3.3.0",
    "openai>=1.3.0",
    "anthropic>=0.8.0",
    "sse-starlette>=1.8.0",
    "websockets>=12.0",
    "cryptography (>=46.0.3,<47.0.0)",
    "dill>=0.3.7",
    "clickhouse-connect>=0.7.0",
    "Jinja2>=3.1.0",
    "aiosmtpd>=1.4.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "N", "B", "A", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]